requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
landingai-ade==0.20.3
psycopg2-binary>=2.9.9
pgvector==0.2.4
google-generativeai==0.3.1
openai>=1.0.0
sentence-transformers>=2.2.0
python-dotenv==1.0.0
pydantic>=2.5.0
numpy>=1.24.3
requests==2.31.0
boto3==1.34.0
PyPDF2==3.0.1

